Replace zoom hook
Replace zoom hook with a smooth interpolation animation hook using expoential smoothing. Change background anchor point to further up the page so zoom focus is on the monitor screen. Increase zoom to 1.75 from 1.5 to show more of the screen. Disable background hook to focus on smooth interpolation hook for now.